[24262번] 알고리즘 수업 - 알고리즘의 수행 시간 1

+++ Beakjoon/01_by_level/11_time_complexity/01_24262.cpp
@@ -4,7 +4,7 @@
 using namespace std;
 
 // 코드 1의 수행 횟수
-int count = 0;
+long long count = 0;
 
 // 알고리즘
 int MenOfPassion(const vector<int>& vec, const int n)
